requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pillow==10.1.0

